+++ example/main.dart
@@ -4,17 +4,17 @@ class ExampleClass {
   LatLongConverter converter = new LatLongConverter();
 
   void OSGBfromDecExample(double lat, double long) {
-    var result = converter.getOSGBfromDec(lat, long);
+    OSRef result = converter.getOSGBfromDec(lat, long);
     print("${result.easting} ${result.northing}");
   }
 
   void OSGBfromDmsExample(var latDeg, var latMin, var latSec, var longDeg, var longMin, var longSec) {
-    var result = converter.getOSGBfromDms(latDeg, latMin, latSec, longDeg, longMin, longSec);
+    OSRef result = converter.getOSGBfromDms(latDeg, latMin, latSec, longDeg, longMin, longSec);
     print("${result.easting} ${result.northing}");
   }
 
   void LatLongfromOSGBExample(var easting, var northing) {
-    var result = converter.getLatLongFromOSGB(easting, northing);
+    LatLong result = converter.getLatLongFromOSGB(easting, northing);
     print("${result.lat} ${result.long}");
   }