+++ lib/maths/CartesianDatum.dart
@@ -1,5 +1,5 @@
 import 'Cartesian.dart';
-import 'LatLong.dart';
+import 'package:latlong_to_osgrid/latlong_to_osgrid.dart';
 import 'package:vector_math/vector_math.dart';
 
 class CartesianDatum extends Cartesian {