+++ test/latlong_to_osgrid_test.dart
@@ -1,13 +1,12 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 import 'package:latlong_to_osgrid/latlong_to_osgrid.dart';
 
 void main() {
   test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
-    expect(() => calculator.addOne(null), throwsNoSuchMethodError);
+    final converter = LatLongConverter();
+    expect(converter.getDecimalFromDegree(52, 57, 36), 52.96);
+    expect(converter.getDecimalFromDegree(1, 4, 48), 1.08);
+    expect(() => converter.getDecimalFromDegree(null, null, null), throwsNoSuchMethodError);
   });
 }