+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.1.0] - 24/08/2020
+
+* Moved LatLong and OSRef objects to main dart file
+* This means you can now declare OSRef and LatLong objects with the one import
+* Updated readme
+
+
+## [1.0.3] - 24/08/2020
+
+* Updated example file and readme
+
+## [1.0.2] - 24/08/2020
+
+* Updated pubspec.yaml - removed flutter dependency
+
 ## [1.0.1] - 24/08/2020
 
 * Updated package description in pubspec.yaml